+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 6.1.0 (IN PROGRESS)
 
+* Add "Mark late" action to the piece form action menu. Refs UIREC-413.
+
 ## [6.0.5](https://github.com/folio-org/ui-receiving/tree/v6.0.5) (2024-12-13)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v6.0.4...v6.0.5)